chore: renamed codeview.py
